DrawnWaveform returns nice lines when zoomed in to the max and should always return an image with a valid span.

+++ tfr/drawnwaveform.h
@@ -19,7 +19,7 @@ public:
 
     virtual FreqAxis freqAxis( float FS );
 
-    unsigned blob(float FS);
+    float blob(float FS);
 
     float block_fs;
     unsigned signal_length;
